added new socket structure
