+++ CMakeLists.txt
@@ -22,9 +22,13 @@ endif ()
 if (APPLE OR BSD)
     list(FILTER UVENT_HEADERS EXCLUDE REGEX ".*/EPoller\\.h$")
     list(FILTER UVENT_SOURCES EXCLUDE REGEX ".*/EPoller\\.cpp$")
+    list(FILTER UVENT_SOURCES EXCLUDE REGEX ".*/SocketLinux\\.h$")
+    list(FILTER UVENT_SOURCES EXCLUDE REGEX ".*/SocketLinux\\.cpp$")
 elseif (UNIX) # Linux
     list(FILTER UVENT_HEADERS EXCLUDE REGEX ".*/KPoller\\.h$")
     list(FILTER UVENT_SOURCES EXCLUDE REGEX ".*/KPoller\\.cpp$")
+    list(FILTER UVENT_SOURCES EXCLUDE REGEX ".*/SocketBSD\\.h$")
+    list(FILTER UVENT_SOURCES EXCLUDE REGEX ".*/SocketBSD\\.cpp$")
 elseif (WIN32)
     message(FATAL_ERROR "Uvent does not support Windows.")
 endif ()
